(hds-1467) Change tests suggested in the PR
